Moved filter_cf.
